real-estate-etl/pyproject.toml
[project]
name = "data-pipelines"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "duckdb>=1.1.3",
    "polars>=1.21.0",
    "pyarrow>=19.0.0",
    "python-dotenv>=1.0.1",
    "python-telegram-bot==13.13",
    "requests>=2.32.3",
]

[tool.uv.workspace]
members = ["real-estate-etl"]
